fixed bugs
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@7857 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/slc/apps/runcontrold/src/RCMasterCallback.cc
@@ -158,7 +158,7 @@ bool RCMasterCallback::trigft() throw()
       if (ftsw->hasValue("trigger_type") && ftsw->hasValue("dummy_rate") &&
           ftsw->hasValue("trigger_limit")) {
         NSMMessage& nsm(getMessage());
-        if (nsm.getNParams() > 2 && nsm.getParam(0) >= 0) {
+        if (nsm.getNParams() > 2 && (int)nsm.getParam(0) >= 0) {
           ftsw->setInt("trigger_type", nsm.getParam(0));
           ftsw->setInt("dummy_rate", nsm.getParam(1));
           ftsw->setInt("trigger_limit", nsm.getParam(2));
